Move stub helper into a separate file.
It will be needed in other test cases so we might as well make it
reusable.
